requirements.txt
# Core Framework
fastapi>=0.104,<1.0
uvicorn[standard]>=0.24,<1.0
pydantic>=2.5,<3.0
pydantic-settings>=2.1,<3.0
python-multipart>=0.0.6,<1.0

# Environment & Configuration
python-dotenv>=1.0,<2.0

# HTTP & Async
httpx>=0.25,<1.0

# Web Scraping
beautifulsoup4>=4.12,<5.0

# Database & Caching
asyncpg>=0.29,<1.0
redis>=5.0,<6.0

# AI & NLP
anthropic>=0.40,<1.0
fuzzywuzzy>=0.18,<1.0
python-Levenshtein>=0.23,<1.0

# Knowledge Graph
neo4j>=5.20,<6.0

# Embeddings (Voyage AI) — uses httpx REST calls, no SDK needed

# Reports & Export
openpyxl>=3.1,<4.0
reportlab>=4.0,<5.0

# Security & Auth
PyJWT>=2.8,<3.0
bcrypt>=4.0,<5.0
cryptography>=42.0,<45.0

# Google APIs (Gmail OAuth2)
google-auth>=2.28,<3.0
google-auth-oauthlib>=1.2,<2.0
google-api-python-client>=2.120,<3.0

# Scheduling
apscheduler>=3.10,<4.0

# Rate Limiting
slowapi>=0.1.9,<1.0

# Monitoring
prometheus-client>=0.19,<1.0

# Development (optional)
pytest>=7.4,<9.0
pytest-asyncio>=0.21,<1.0
black>=23.11
flake8>=6.1
